Remove lifetime from parquet reader
Refactor module structure
