+++ nautilus_core/common/Cargo.toml
@@ -14,7 +14,6 @@ nautilus_core = { path = "../core" }
 nautilus_model = { path = "../model" }
 pyo3 = { version = "0.16.5" }
 arrow2 = { version = "0.12.0", features = [ "io_parquet", "io_csv_read" ] }
-chrono = "^0.4.19"
 
 [features]
 extension-module = ["pyo3/extension-module"]
